tests/wpt/tests/html/canvas/offscreen/transformations/2d.transformation.order.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.transformation.order</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.transformation.order</h1>
<p class="desc">Transformations are applied in the right order</p>


<script>
var t = async_test("Transformations are applied in the right order");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);

  ctx.scale(2, 1);
  ctx.rotate(Math.PI / 2);
  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, -50, 50, 50);
  _assertPixel(canvas, 75,25, 0,255,0,255);
  t.done();

});
</script>
